[routing] review fixes

+++ routing/cross_mwm_connector.hpp
@@ -151,7 +151,7 @@ public:
     auto const & transition = GetTransition(segment);
     CHECK_NOT_EQUAL(transition.m_enterIdx, connector::kFakeIndex, ());
     auto const enterIdx = transition.m_enterIdx;
-    for (uint32_t exitIdx = 0; exitIdx < static_cast<uint32_t>(m_exits.size()); ++exitIdx)
+    for (size_t exitIdx = 0; exitIdx < m_exits.size(); ++exitIdx)
     {
       auto const weight = GetWeight(enterIdx, exitIdx);
       AddEdge(m_exits[exitIdx], weight, edges);
@@ -163,7 +163,7 @@ public:
     auto const & transition = GetTransition(segment);
     CHECK_NOT_EQUAL(transition.m_exitIdx, connector::kFakeIndex, ());
     auto const exitIdx = transition.m_exitIdx;
-    for (uint32_t enterIdx = 0; enterIdx < static_cast<uint32_t>(m_enters.size()); ++enterIdx)
+    for (size_t enterIdx = 0; enterIdx < m_enters.size(); ++enterIdx)
     {
       auto const weight = GetWeight(enterIdx, exitIdx);
       AddEdge(m_enters[enterIdx], weight, edges);
